ポーズ画面のボタンが点滅するようにしました（白石）

+++ ShootingGameFramework/ShootingGame/src/canTouch/GameObject/PauseObject/RestartButton/RestartButtonScript.cpp
@@ -5,8 +5,32 @@ RestartButtonScript::RestartButtonScript()
 }
 
 void RestartButtonScript::update()
-{
-	if (Keyboard::getState(InputType::INPUT_PUSHING, KeyboardKeyType::KEYBOARD_SPACE)||GamePad::getState(GamePadNumber::GAME_PAD_NUMBER_1,InputType::INPUT_BEGIN,GamePadButtonType::GAME_PAD_B_BUTTON))
+{ 
+	counter++;
+
+	if (select == 1)
+	{
+		if (counter % 40 == 0)
+		{
+			getComponent<Sprite2dDrawer>().lock()->setActive(false);
+		}
+		else if (counter % 40 == 20)
+		{
+			getComponent<Sprite2dDrawer>().lock()->setActive(true);
+		}
+	}
+
+	if (Keyboard::getState(InputType::INPUT_PUSHING, KeyboardKeyType::KEYBOARD_UP))
+	{
+		select = 1;
+	}
+	else if (Keyboard::getState(InputType::INPUT_PUSHING, KeyboardKeyType::KEYBOARD_DOWN))
+	{
+		select = 2;
+		getComponent<Sprite2dDrawer>().lock()->setActive(true);
+	}
+
+	if (select == 1 && Keyboard::getState(InputType::INPUT_PUSHING, KeyboardKeyType::KEYBOARD_SPACE)||GamePad::getState(GamePadNumber::GAME_PAD_NUMBER_1,InputType::INPUT_BEGIN,GamePadButtonType::GAME_PAD_B_BUTTON))
 	{
 		GameObjectManager::findGameObjectWithTag(GAME_OBJECT_TAG_PLAYER).lock()->destroy();
 		GameObjectManager::sendMessage(DIE_GAMEPLAY_OBJECT);
